refactor useTranspose
